requirements.txt
flask==3.0.3
mysql-connector-python==8.4.0
flask-swagger-ui==4.11.1
python-dotenv==1.0.1
gunicorn==22.0.0
